requirements.txt
python-dotenv~=1.0.1
aiogram~=3.10.0
redis~=5.0.7
aiohttp~=3.9.5